moves groups to constants
